server/requirements.txt
Flask
psycopg2-binary
google-cloud-pubsub